feat: table column resize
